+++ src/main/java/ru/egarcourses/HospitalInfoSystem/controllers/rest/PatientRESTController.java
@@ -9,8 +9,8 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 import ru.egarcourses.HospitalInfoSystem.dto.PatientDTO;
 import ru.egarcourses.HospitalInfoSystem.services.impl.PatientServiceImpl;
-import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotCreatedException;
-import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotUpdatedException;
+import ru.egarcourses.HospitalInfoSystem.utils.exceptions.NotCreatedException;
+import ru.egarcourses.HospitalInfoSystem.utils.exceptions.NotUpdatedException;
 
 import java.util.List;
 
@@ -26,20 +26,20 @@ public class PatientRESTController {
     }
 
     @GetMapping()
-    public ResponseEntity<List<PatientDTO>> index(){
+    public ResponseEntity<List<PatientDTO>> index() {
         final List<PatientDTO> patientDTOList = patientServiceImpl.findAll();
         return ResponseEntity.ok(patientDTOList);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<PatientDTO> show(@PathVariable("id") int id){
+    public ResponseEntity<PatientDTO> show(@PathVariable("id") Long id) {
         final PatientDTO patientDTO = patientServiceImpl.findById(id);
         return ResponseEntity.ok(patientDTO);
     }
 
     @PostMapping()
-    public ResponseEntity<HttpStatus> create(@RequestBody @Valid PatientDTO patientDTO, BindingResult bindingResult){
-        if(bindingResult.hasErrors()){
+    public ResponseEntity<HttpStatus> create(@RequestBody @Valid PatientDTO patientDTO, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
             StringBuilder errorMessage = new StringBuilder();
             List<FieldError> fieldErrors = bindingResult.getFieldErrors();
             for (FieldError fieldError : fieldErrors) {
@@ -54,8 +54,8 @@ public class PatientRESTController {
 
     @PatchMapping("/{id}")
     public ResponseEntity<HttpStatus> update(@RequestBody @Valid PatientDTO patientDTO, BindingResult bindingResult,
-                         @PathVariable("id") int id){
-        if(bindingResult.hasErrors()){
+                                             @PathVariable("id") Long id) {
+        if (bindingResult.hasErrors()) {
             StringBuilder errorMessage = new StringBuilder();
             List<FieldError> fieldErrors = bindingResult.getFieldErrors();
             for (FieldError fieldError : fieldErrors) {
@@ -69,7 +69,7 @@ public class PatientRESTController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> delete(@PathVariable("id") int id){
+    public ResponseEntity<HttpStatus> delete(@PathVariable("id") Long id) {
         patientServiceImpl.delete(id);
         return ResponseEntity.ok(HttpStatus.OK);
     }
